requirements.txt
PyYAML==6.0.1
Flask==3.0.0
